third Init

+++ helloworld.go
@@ -13,7 +13,6 @@ func main() {
 	var tax float64 = taxRate * float64(price)
 
 	fmt.Println("Tax is", tax, "dollars")
-
-	var total float64 = float64(price) + float64(tax)
+	var total float64 = float64(price) + tax
 	fmt.Println("Total cost is", total, "dollars")
 }
